--- src/main/java/com/example/schedule/board/repository/BoardRepository.java
@@ -1,12 +1,0 @@
-package com.example.schedule.board.repository;
-
-import com.example.schedule.board.entity.Board;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.server.ResponseStatusException;
-
-public interface BoardRepository extends JpaRepository<Board,Long> {
-    default Board findByIdOrElseThrow(Long id){
-        return findById(id).orElseThrow(()-> new ResponseStatusException(HttpStatus.NOT_FOUND,  "Does not exist id = " + id));
-    }
-}
