+++ src/main/java/com/example/schedule/common/Const.java
@@ -1,0 +1,5 @@
+package com.example.schedule.common;
+
+public interface Const {
+    String LOGIN_USER ="LoginUser";
+}
