+++ src/main/java/com/example/schedule/config/WebConfig.java
@@ -1,0 +1,22 @@
+package com.example.schedule.config;
+
+import com.example.schedule.filter.LoginFilter;
+import org.springframework.boot.web.servlet.FilterRegistrationBean;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+
+@Configuration
+public class WebConfig implements WebMvcConfigurer {
+
+    @Bean
+    public FilterRegistrationBean loginFilter() {
+        FilterRegistrationBean<LoginFilter> filterRegistrationBean = new FilterRegistrationBean<>();
+        filterRegistrationBean.setFilter(new LoginFilter()); // Filter 등록
+        filterRegistrationBean.setOrder(2); // Filter 순서 2 설정
+        filterRegistrationBean.addUrlPatterns("/*"); // 전체 URL에 Filter 적용
+
+        return filterRegistrationBean;
+    }
+}
