+++ src/main/java/com/example/schedule/domain/members/dto/UpdatePasswordRequestDto.java
@@ -1,0 +1,14 @@
+package com.example.schedule.domain.members.dto;
+
+import lombok.Getter;
+
+@Getter
+public class UpdatePasswordRequestDto {
+    private final String oldPassword;
+    private final String newPassword;
+
+    public UpdatePasswordRequestDto(String oldPassword, String newPassword){
+        this.oldPassword = oldPassword;
+        this.newPassword = newPassword;
+    }
+}
