+++ src/main/java/com/example/schedule/domain/members/dto/login/LoginResponseDto.java
@@ -1,0 +1,12 @@
+package com.example.schedule.domain.members.dto.login;
+
+import lombok.Getter;
+
+@Getter
+public class LoginResponseDto {
+    private final Long id;
+
+    public LoginResponseDto(Long id) {
+        this.id = id;
+    }
+}
