+++ src/main/java/com/example/schedule/exception/CustomException.java
@@ -1,0 +1,9 @@
+package com.example.schedule.exception;
+
+import org.springframework.http.HttpStatus;
+
+public class CustomException extends RuntimeException {
+    public CustomException(String message) {
+        super(message);
+    }
+}
