+++ src/main/java/com/example/schedule/exception/GlobalExceptionHandler.java
@@ -1,9 +1,39 @@
 package com.example.schedule.exception;
 
 
+import org.hibernate.dialect.unique.CreateTableUniqueDelegate;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.validation.FieldError;
+import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
+import org.springframework.web.client.HttpClientErrorException;
+import org.springframework.web.client.RestClientException;
+import org.springframework.web.server.ResponseStatusException;
 
-@RestControllerAdvice
+import java.util.HashMap;
+import java.util.Map;
+
+@ControllerAdvice
 public class GlobalExceptionHandler {
 
+    @ExceptionHandler(IllegalArgumentException.class)
+    public ResponseEntity<String> handleIllegalArgumentException(IllegalArgumentException ex){
+        return ResponseEntity.badRequest().body("Global Error" + ex.getMessage());
+    }
+
+    @ExceptionHandler(MethodArgumentNotValidException.class)
+    public ResponseEntity<Map<String, String>> handleMethodArqumentValidException(MethodArgumentNotValidException ex){
+        Map<String, String> errors = new HashMap<>();
+        for(FieldError fieldError : ex.getBindingResult().getFieldErrors()){
+            errors.put(fieldError.getField(), fieldError.getDefaultMessage());
+        }
+        return new ResponseEntity<>(errors, HttpStatus.INTERNAL_SERVER_ERROR);
+    }
+    @ExceptionHandler(ResponseStatusException.class)
+    private ResponseEntity<String> handleResponseStatusException(ResponseStatusException ex){
+        return  new ResponseEntity<>(ex.getMessage(), HttpStatus.NOT_FOUND);
+    }
 }
