--- src/main/java/com/example/schedule/members/dto/UserResponseDto.java
@@ -1,15 +1,0 @@
-package com.example.schedule.members.dto;
-
-import lombok.Getter;
-
-@Getter
-public class UserResponseDto {
-    private final Long id;
-    private final String name;
-
-
-    public UserResponseDto(Long id, String name) {
-        this.id = id;
-        this.name = name;
-    }
-}
