--- src/main/java/com/example/schedule/members/dto/sign/SignUpRequestDto.java
@@ -1,18 +1,0 @@
-package com.example.schedule.members.dto.sign;
-
-import lombok.Getter;
-
-@Getter
-public class SignUpRequestDto {
-    private final String userName;
-
-    private final String password;
-
-    private final String email;
-
-    public SignUpRequestDto(String userName, String password, String email) {
-        this.userName = userName;
-        this.password = password;
-        this.email = email;
-    }
-}
