--- src/main/java/com/example/schedule/members/dto/sign/SignUpResponseDto.java
@@ -1,15 +1,0 @@
-package com.example.schedule.members.dto.sign;
-
-import lombok.Getter;
-
-@Getter
-public class SignUpResponseDto {
-    private final Long id;
-
-    private final String userName;
-
-    public SignUpResponseDto(Long id, String userName) {
-        this.id = id;
-        this.userName = userName;
-    }
-}
